* Passage par des value Long * Refactoring du code

+++ app/src/main/java/com/fizzbuzz/data/repository/FizzbuzzPagingSource.kt
@@ -7,24 +7,24 @@ import com.fizzbuzz.model.FizzbuzzEntity
 import javax.inject.Inject
 
 
-fun Int.maximize(limit: Int): Int = if (this < limit) this else limit
+fun Long.maximize(limit: Long): Long = if (this < limit) this else limit
 
 class FizzbuzzPagingSource @Inject
 constructor(
     val entity: FizzbuzzEntity,
     private val getFizzbuzzStringUseCase: GetFizzbuzzStringUseCase
 ) :
-    PagingSource<Int, String>() {
+    PagingSource<Long, String>() {
 
     companion object {
         const val PAGE_SIZE = 100
     }
 
-    override suspend fun load(params: LoadParams<Int>):
-            LoadResult<Int, String> {
+    override suspend fun load(params: LoadParams<Long>):
+            LoadResult<Long, String> {
         val currentLoadingPageKey = params.key ?: 0
         val res: ArrayList<String> = ArrayList()
-        val prevKey = if (currentLoadingPageKey == 0) null else currentLoadingPageKey - 1
+        val prevKey = if (currentLoadingPageKey == 0.toLong()) null else currentLoadingPageKey - 1
 
         val currentIndex = (1 + (currentLoadingPageKey) * PAGE_SIZE)
         return if (currentIndex < entity.limit) {
@@ -42,13 +42,12 @@ constructor(
         }
     }
 
-    override fun getRefreshKey(state: PagingState<Int, String>): Int? {
+    override fun getRefreshKey(state: PagingState<Long, String>): Long? {
 
         return state.anchorPosition?.let { anchorPosition ->
             val anchorPage = state.closestPageToPosition(anchorPosition)
             anchorPage?.prevKey?.plus(1) ?: anchorPage?.nextKey?.minus(1)
         }
     }
-
 
 }
