+++ app/src/main/java/com/fizzbuzz/domain/GetFizzbuzzPagerUseCase.kt
@@ -5,7 +5,7 @@ import javax.inject.Inject
 
 class GetFizzbuzzPagerUseCase @Inject
 constructor(
-    private val pager: Pager<Int, String>
+    private val pager: Pager<Long, String>
 ) {
     operator fun invoke() = pager.flow
 }
