+++ app/src/main/java/com/fizzbuzz/domain/GetFizzbuzzStringUseCase.kt
@@ -12,17 +12,16 @@ constructor() {
      * @property item the fizzbuzz parameters.
      * @return The fizzbuzz string
      */
-    fun invoke(index: Int, item: FizzbuzzEntity): String {
+    fun invoke(index: Long, item: FizzbuzzEntity): String {
         var str = ""
         // All multiples of int1 are replaced by str1
-        if (index % item.int1 == 0) {
+        if (index % item.int1 == 0.toLong()) {
             str += item.str1
         }
         // All multiples of int2 are replaced by str2
-        if (index % item.int2 == 0) {
+        if (index % item.int2 == 0.toLong()) {
             str += item.str2
         }
         return str.ifEmpty { index.toString() }
     }
-
 }
